requirements(1).txt
discord.py
mysql-connector-python
pytz
openai
yt_dlp
PyNaCl